add localized paths to project
