+++ server/src/compiler/utilitiesPublic.ts
@@ -1445,3 +1445,4 @@ export function validateLocaleAndSetLanguage(
     //     return true;
     // }
 }
+
